Update Streamr destination to handle stream name behaviors
This commit introduces changes to the Streamr destination's handling of stream names. It allows the configuration of different behaviors such as using the Airbyte stream name as the suffix, setting the stream name as a data property, or ignoring the name. Also, the "streamAddress" property is replaced with "streamrStreamPrefix", offering more flexibility in constructing stream names.